Add possibility of using the long_title as directory name of game
Projects like GameHub use the long title of the game as the name of the
directory that stores the installers. If gogrepo used the same structure we
could use the gogrepo offline backup directly as the cache of game installers in
GameHub, and in this way save time and bandwidth on needless downloads.

This option is off by default for compatibility and because the (shorter) titles
are a bit better supported in filesystems (and consistent?) than the longer ones.