Add the possibility of storing MD5 hashes along with the files
GameHub makes use of the MD5 hashes and stores them alongside the installers.
If we were to use gogrepo's backups as the source of installers for GameHub,
having the MD5 hashes already present would be convenient.

An option for storing said hashes as files is present in the code, and disabled
by default.